Moved menu commands to MenuController
New sub-controller.
