+++ src/aslm/controller/controller.py
@@ -34,33 +34,20 @@
 #  Standard Library Imports
 from multiprocessing import Manager
 import tkinter
-from tkinter import filedialog, messagebox
+from tkinter import messagebox
 import multiprocessing as mp
 import threading
 import sys
-import platform
 
 # Third Party Imports
 
 # Local View Imports
 from aslm.view.main_application_window import MainApp as view
-from aslm.view.popups.waveform_parameter_popup_window import (
-    WaveformParameterPopupWindow,
-)
 from aslm.view.popups.camera_view_popup_window import CameraViewPopupWindow
-from aslm.view.popups.autofocus_setting_popup import AutofocusPopup
-from aslm.view.popups.ilastik_setting_popup import ilastik_setting_popup
-from aslm.view.popups.help_popup import HelpPopup
-from aslm.view.popups.camera_map_setting_popup import CameraMapSettingPopup
 
 # Local Sub-Controller Imports
-from aslm.controller.sub_controllers.help_popup_controller import HelpPopupController
 from aslm.controller.configuration_controller import ConfigurationController
 from aslm.controller.sub_controllers import (
-    IlastikPopupController,
-    CameraMapSettingPopupController,
-    AutofocusPopupController,
-    WaveformPopupController,
     KeystrokeController,
     WaveformTabController,
     StageController,
@@ -69,7 +56,7 @@ from aslm.controller.sub_controllers import (
     MultiPositionController,
     ChannelsTabController,
     AcquireBarController,
-    MicroscopePopupController,
+    MenuController,
 )
 from aslm.controller.thread_pool import SynchronizedThreadPool
 
@@ -180,43 +167,30 @@ class Controller:
         self.view.root.protocol("WM_DELETE_WINDOW", self.exit_program)
 
         # Sub Gui Controllers
-        # Acquire bar, channels controller,
-        # camera view, camera settings,
-        # stage, waveforms, menus.
         self.acquire_bar_controller = AcquireBarController(
             self.view.acqbar, self.view.settings.channels_tab, self
         )
-
         self.channels_tab_controller = ChannelsTabController(
             self.view.settings.channels_tab, self
         )
-
         self.multiposition_tab_controller = MultiPositionController(
             self.view.settings.multiposition_tab.multipoint_list, self
         )
-
         self.camera_view_controller = CameraViewController(
             self.view.camera_waveform.camera_tab, self
         )
-
         self.camera_setting_controller = CameraSettingController(
             self.view.settings.camera_settings_tab, self
         )
-
-        # Stage Controller
         self.stage_controller = StageController(
             self.view.settings.stage_control_tab,
             self.view,
             self.camera_view_controller.canvas,
             self,
         )
-
-        # Waveform Controller
         self.waveform_tab_controller = WaveformTabController(
             self.view.camera_waveform.waveform_tab, self
         )
-
-        # Keystroke Controller
         self.keystroke_controller = KeystrokeController(self.view, self)
 
         # Bonus config
@@ -228,7 +202,9 @@ class Controller:
         # self.microscope = self.configuration['configuration']
         # ['microscopes'].keys()[0]  # Default to the first microscope
 
-        self.initialize_menus(args.synthetic_hardware)
+        # Initialize the menus
+        self.menu_controller = MenuController(view=self.view, parent_controller=self)
+        self.menu_controller.initialize_menus(args.synthetic_hardware)
 
         # Create default data buffer
         self.img_width = 0
@@ -315,340 +291,6 @@ class Controller:
         image_metrics = [1, 0, 0]
         self.camera_view_controller.initialize("image", image_metrics)
 
-    def initialize_menus(self, is_synthetic_hardware=False):
-        """Initialize menus
-        This function defines all the menus in the menubar
-
-        Parameters
-        ----------
-        is_synthetic_hardware : bool
-            If True, then the hardware is simulated.
-            If False, then the hardware is real.
-
-        Returns
-        -------
-        configuration_controller : class
-            Camera view sub-controller.
-
-        """
-
-        def new_experiment(*args):
-            """Create a new experiment file."""
-            self.populate_experiment_setting(self.default_experiment_file)
-
-        def load_experiment(*args):
-            """Load an experiment file."""
-            filename = filedialog.askopenfilename(
-                defaultextension=".yml", filetypes=[("Yaml files", "*.yml *.yaml")]
-            )
-            if not filename:
-                return
-            self.populate_experiment_setting(filename)
-
-        def save_experiment(*args):
-            """Save an experiment file.
-
-            Updates model.experiment and saves it to file.
-            """
-            if not self.update_experiment_setting():
-                tkinter.messagebox.showerror(
-                    title="Warning",
-                    message="Incorrect/missing settings. "
-                    "Cannot save current experiment file.",
-                )
-                return
-            filename = filedialog.asksaveasfilename(
-                defaultextension=".yml", filetypes=[("Yaml file", "*.yml *.yaml")]
-            )
-            if not filename:
-                return
-            save_yaml_file("", self.configuration["experiment"], filename)
-
-        def load_images():
-            """Load images from a file."""
-            filenames = filedialog.askopenfilenames(
-                defaultextension=".tif", filetypes=[("tiff files", "*.tif *.tiff")]
-            )
-            if not filenames:
-                return
-            self.model.load_images(filenames)
-
-        def popup_waveform_setting():
-            if hasattr(self, "waveform_popup_controller"):
-                self.waveform_popup_controller.showup()
-                return
-            waveform_constants_popup = WaveformParameterPopupWindow(
-                self.view, self.configuration_controller
-            )
-            self.waveform_popup_controller = WaveformPopupController(
-                waveform_constants_popup, self, self.waveform_constants_path
-            )
-
-            self.waveform_popup_controller.populate_experiment_values()
-
-        def popup_microscope_setting():
-            """Pop up the microscope setting window.
-
-            Parameters
-            ----------
-            None
-
-            Returns
-            -------
-            None
-            """
-            if hasattr(self, "microscope_popup_controller"):
-                self.microscope_popup_controller.showup()
-                return
-            microscope_info = self.model.get_microscope_info()
-            self.microscope_popup_controller = MicroscopePopupController(
-                self.view, self, microscope_info
-            )
-
-        def popup_autofocus_setting(*args):
-            """Pop up the Autofocus setting window."""
-            if hasattr(self, "af_popup_controller"):
-                self.af_popup_controller.showup()
-                return
-            af_popup = AutofocusPopup(self.view)
-            self.af_popup_controller = AutofocusPopupController(af_popup, self)
-
-        def popup_camera_map_setting():
-            """Pop up the Camera Map setting window."""
-            if hasattr(self, "camera_map_popup_controller"):
-                self.camera_map_popup_controller.showup()
-                return
-            map_popup = CameraMapSettingPopup(self.view)
-            self.camera_map_popup_controller = CameraMapSettingPopupController(
-                map_popup, self
-            )
-
-        def popup_ilastik_setting():
-            """Pop up the Ilastik setting window."""
-            ilastik_popup_window = ilastik_setting_popup(self.view)
-            ilastik_url = self.configuration["rest_api_config"]["Ilastik"]["url"]
-            if hasattr(self, "ilastik_controller"):
-                self.ilastik_controller.showup(ilastik_popup_window)
-            else:
-                self.ilastik_controller = IlastikPopupController(
-                    ilastik_popup_window, self, ilastik_url
-                )
-
-        def popup_help():
-            """Pop up the help window."""
-            if hasattr(self, "help_controller"):
-                self.help_controller.showup()
-                return
-            help_pop = HelpPopup(self.view)
-            self.help_controller = HelpPopupController(help_pop, self)
-
-        menus_dict = {
-            self.view.menubar.menu_file: {
-                "New Experiment": [
-                    new_experiment,
-                    "Ctrl+Shift+N",
-                    "<Control-N>",
-                    "<Control_L-N>",
-                ],
-                "Load Experiment": [
-                    load_experiment,
-                    "Ctrl+Shift+O",
-                    "<Control-O>",
-                    "<Control_L-O>",
-                ],
-                "Save Experiment": [
-                    save_experiment,
-                    "Ctrl+Shift+S",
-                    "<Control-S>",
-                    "<Control_L-S>",
-                ],
-                "add_separator": [None],
-                "Save Data": [None, "Ctrl+s", "<Control-s>", "<Control_L-s>"],
-            },
-            self.view.menubar.menu_multi_positions: {
-                "Move Up (X)": [None, "w", "<Key-w>", "<Key-w>"],
-                "Move Down (X)": [None, "s", "<Key-s>", "<Key-s>"],
-                "Move Left (Y)": [None, "a", "<Key-a>", "<Key-a>"],
-                "Move Right (Y)": [None, "d", "<Key-d>", "<Key-d>"],
-                "Move In (Z)": [None, "q", "<Key-q>", "<Key-q>"],
-                "Move Out (Z)": [None, "e", "<Key-e>", "<Key-e>"],
-                "Move Focus Up (F)": [None, "r", "<Key-r>", "<Key-r>"],
-                "Move Focus Down (F)": [None, "f", "<Key-f>", "<Key-f>"],
-                "Rotate Clockwise (R)": [None, "z", "<Key-z>", "<Key-z>"],
-                "Rotate Counter-Clockwise (R)": [None, "x", "<Key-x>", "<Key-x>"],
-                "add_separator": [None],
-                "Launch Tiling Wizard": [None, None, None, None],
-                "Load Positions": [
-                    self.multiposition_tab_controller.load_positions,
-                    None,
-                    None,
-                    None,
-                ],
-                "Export Positions": [
-                    self.multiposition_tab_controller.export_positions,
-                    None,
-                    None,
-                    None,
-                ],
-                "Append Current Position": [
-                    self.multiposition_tab_controller.add_stage_position,
-                    None,
-                    None,
-                    None,
-                ],
-                "Generate Positions": [
-                    self.multiposition_tab_controller.generate_positions,
-                    None,
-                    None,
-                    None,
-                ],
-                "Move to Selected Position": [
-                    self.multiposition_tab_controller.move_to_position,
-                    None,
-                    None,
-                    None,
-                ],
-            },
-        }
-        for menu in menus_dict:
-            menu_items = menus_dict[menu]
-            for label in menu_items:
-                if label == "add_separator":
-                    menu.add_separator()
-                else:
-                    menu.add_command(
-                        label=label,
-                        command=menu_items[label][0],
-                        accelerator=menu_items[label][1],
-                    )
-                    if platform.platform() == "Darwin":
-                        menu.bind_all(menu_items[label][3], menu_items[label][0])
-                    else:
-                        menu.bind_all(menu_items[label][2], menu_items[label][0])
-
-        # load images from disk in synthetic hardware
-        if is_synthetic_hardware:
-            self.view.menubar.menu_file.add_separator()
-            self.view.menubar.menu_file.add_command(
-                label="Load Images", command=load_images
-            )
-            self.view.menubar.menu_file.add_command(
-                label="Unload Images", command=lambda: self.model.load_images(None)
-            )
-
-        # add resolution menu
-        self.resolution_value = tkinter.StringVar()
-        for microscope_name in self.configuration["configuration"][
-            "microscopes"
-        ].keys():
-            zoom_positions = self.configuration["configuration"]["microscopes"][
-                microscope_name
-            ]["zoom"]["position"]
-            if len(zoom_positions) > 1:
-                sub_menu = tkinter.Menu(self.view.menubar.menu_resolution)
-                self.view.menubar.menu_resolution.add_cascade(
-                    menu=sub_menu, label=microscope_name
-                )
-                for res in zoom_positions.keys():
-                    sub_menu.add_radiobutton(
-                        label=res,
-                        variable=self.resolution_value,
-                        value=f"{microscope_name} {res}",
-                    )
-            else:
-                self.view.menubar.menu_resolution.add_radiobutton(
-                    label=microscope_name,
-                    variable=self.resolution_value,
-                    value=f"{microscope_name} {zoom_positions.keys()[0]}",
-                )
-
-        # event binding
-        self.resolution_value.trace_add(
-            "write",
-            lambda *args: self.execute("resolution", self.resolution_value.get()),
-        )
-
-        # add separator
-        self.view.menubar.menu_resolution.add_separator()
-
-        # waveform popup
-        self.view.menubar.menu_resolution.add_command(
-            label="Waveform Parameters", command=popup_waveform_setting
-        )
-        # microscope setting popup
-        self.view.menubar.menu_resolution.add_command(
-            label="Configure Microscopes", command=popup_microscope_setting
-        )
-
-        # autofocus menu
-        self.view.menubar.menu_autofocus.add_command(
-            label="Perform Autofocus",
-            command=lambda: self.execute("autofocus"),
-            accelerator="Ctrl+A",
-        )
-        self.view.menubar.menu_autofocus.add_command(
-            label="Autofocus Settings",
-            command=popup_autofocus_setting,
-            accelerator="Ctrl+Shift+A",
-        )
-
-        if platform.platform == "darwin":
-            self.view.bind_all("<Control_L-a>", lambda event: self.execute("autofocus"))
-            self.view.bind_all("<Control_L-A>", popup_autofocus_setting)
-        else:
-            self.view.bind_all("<Control-a>", lambda event: self.execute("autofocus"))
-            self.view.bind_all("<Control-A>", popup_autofocus_setting)
-
-        # Window Menu
-        self.view.menubar.menu_window.add_command(
-            label="Channel Settings", accelerator="Ctrl+1"
-        )
-        self.view.menubar.menu_window.add_command(
-            label="Camera Settings", accelerator="Ctrl+2"
-        )
-        self.view.menubar.menu_window.add_command(
-            label="Stage Control", accelerator="Ctrl+3"
-        )
-        self.view.menubar.menu_window.add_command(
-            label="Multiposition Table", accelerator="Ctrl+4"
-        )
-        self.view.menubar.menu_window.add_separator()
-        self.view.menubar.menu_window.add_command(label="Popout Camera Display")
-
-        # Help menu
-        self.view.menubar.menu_help.add_command(label="Help", command=popup_help)
-
-        # add-on features
-        feature_list = [
-            "None",
-            "Switch Resolution",
-            "Z Stack Acquisition",
-            "Threshold",
-            "Ilastik Segmentation",
-            "Volume Search",
-            "Time Series",
-        ]
-        self.feature_id_val = tkinter.IntVar(0)
-        for i in range(len(feature_list)):
-            self.view.menubar.menu_features.add_radiobutton(
-                label=feature_list[i], variable=self.feature_id_val, value=i
-            )
-        self.feature_id_val.trace_add(
-            "write",
-            lambda *args: self.execute("load_feature", self.feature_id_val.get()),
-        )
-        self.view.menubar.menu_features.add_separator()
-        self.view.menubar.menu_features.add_command(
-            label="Ilastik Settings", command=popup_ilastik_setting
-        )
-        # disable ilastik menu
-        self.view.menubar.menu_features.entryconfig(
-            "Ilastik Segmentation", state="disabled"
-        )
-        self.view.menubar.menu_features.add_command(
-            label="Camera offset and variance maps", command=popup_camera_map_setting
-        )
-
     def populate_experiment_setting(self, file_name=None, in_initialize=False):
         """Load experiment file and populate model.experiment and configure view.
 
@@ -672,7 +314,7 @@ class Controller:
         microscope_name = self.configuration["experiment"]["MicroscopeState"][
             "microscope_name"
         ]
-        self.resolution_value.set(
+        self.menu_controller.resolution_value.set(
             f"{microscope_name} "
             f"{self.configuration['experiment']['MicroscopeState']['zoom']}"
         )
@@ -696,9 +338,8 @@ class Controller:
     def update_experiment_setting(self):
         """Update model.experiment according to values in the GUI
 
-        Collect settings from sub-controllers
-        sub-controllers will validate the value, if something is wrong, it will
-        return False
+        Collect settings from sub-controllers will validate the value, if something
+        is wrong, it will return False
 
         """
         # acquire_bar_controller - update image mode
@@ -763,7 +404,7 @@ class Controller:
         if mode == "stop":
             # GUI Failsafe
             self.acquire_bar_controller.stop_acquire()
-            self.feature_id_val.set(0)
+            self.menu_controller.feature_id_val.set(0)
 
     def execute(self, command, *args):
         """Functions listens to the Sub_Gui_Controllers.
