+++ src/aslm/controller/sub_controllers/__init__.py
@@ -1,13 +1,14 @@
-from .stage_controller import StageController
-from .acquire_bar_controller import AcquireBarController
-from .channels_tab_controller import ChannelsTabController
-from .camera_view_controller import CameraViewController
-from .camera_setting_controller import CameraSettingController
-from .waveform_tab_controller import WaveformTabController
-from .waveform_popup_controller import WaveformPopupController
-from .autofocus_popup_controller import AutofocusPopupController
-from .keystroke_controller import KeystrokeController
-from .multi_position_controller import MultiPositionController
-from .ilastik_popup_controller import IlastikPopupController
-from .camera_map_setting_popup_controller import CameraMapSettingPopupController
-from .microscope_popup_controller import MicroscopePopupController
+from .stage_controller import StageController  # noqa
+from .acquire_bar_controller import AcquireBarController  # noqa
+from .channels_tab_controller import ChannelsTabController  # noqa
+from .camera_view_controller import CameraViewController  # noqa
+from .camera_setting_controller import CameraSettingController  # noqa
+from .waveform_tab_controller import WaveformTabController  # noqa
+from .waveform_popup_controller import WaveformPopupController  # noqa
+from .autofocus_popup_controller import AutofocusPopupController  # noqa
+from .keystroke_controller import KeystrokeController  # noqa
+from .multi_position_controller import MultiPositionController  # noqa
+from .ilastik_popup_controller import IlastikPopupController  # noqa
+from .camera_map_setting_popup_controller import CameraMapSettingPopupController  # noqa
+from .microscope_popup_controller import MicroscopePopupController  # noqa
+from .menu_controller import MenuController  # noqa
